% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10int.R
\name{log10int.default}
\alias{log10int.default}
\title{Interpolated log10}
\usage{
\method{log10int}{default}(x, ...)
}
\arguments{
\item{x}{a numeric or vector of numerics}

\item{...}{further arguments passed to or from other methods}
}
\value{
a warning that the method cannot handle the given class of data
}
\description{
This is the default mwthod of log10int. It prints a warning if an object of another type than numeric
is given to the function.
}
