% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10int.R
\name{log10int}
\alias{log10int}
\title{Interpolated log10}
\usage{
log10int(x, ...)
}
\arguments{
\item{x}{a numeric or vector of numerics}

\item{...}{further arguments passed to or from other methods}
}
\value{
numeric or vector of numerics, containing the interpolated log10 values of x.
If no object of the class numeric is given to the function, a warning is printed.
}
\description{
log10int interpolates the log10 between 1 and 100. Therefor it uses a table with the log10 values
between 1 and 100 in steps of 0.01, computed with the R log10 function. The accuracy of the interpolation
is in the order of 1e-6 in mean and in the order of 1e-3 in the worst case (tested on the example column in
log10data, which consists of randomly generated numbers between 1 and 100).
It is a generic function and S3 methods exists for the class numeric and default.
}
\examples{
log10int(5)
log10int(c(3.77, 6.228))

}
