% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{log10data}
\alias{log10data}
\title{Values of log10 to interpolate between}
\format{A dataframe with 9901 rows and 3 columns.}
\usage{
log10data
}
\description{
The data is generated with R, such that the first column are the values from
1 to 100 in steps of 0.01. The second column is the log10 function from R of
the first column. The third column contains randomly generated numbers, sorted,
between 1 and 100 to test the log10int method.
}
\keyword{datasets}
